Ignore symlinks.
Especially important if we want to ensure that the user may want to invoke the
process again later, if they are keeping multiple envs with different
versions.

+++ src/diffusers/utils/hub_utils.py
@@ -152,7 +152,7 @@ def move_cache(old_cache_dir: Optional[str] = None, new_cache_dir: Optional[str]
     old_cache_dir = Path(old_cache_dir).expanduser()
     new_cache_dir = Path(new_cache_dir).expanduser()
     for old_blob_path in old_cache_dir.glob("**/blobs/*"): #  move file blob by blob
-        if old_blob_path.is_file():
+        if old_blob_path.is_file() and not old_blob_path.is_symlink():
             new_blob_path = new_cache_dir / old_blob_path.relative_to(old_cache_dir)
             new_blob_path.parent.mkdir(parents=True, exist_ok=True)
             os.replace(old_blob_path, new_blob_path)
